+++ src/app/(main)/profile/[userId]/components/ProfileByIdDetail.client.tsx
@@ -1,0 +1,12 @@
+'use client';
+import ProfileDetailSection from '../../components/ProfileDetailSection';
+import { useGetProfileById } from '@/apis/profile';
+import { useNumberParams } from '@/hooks/useNumberParams';
+
+export default function ProfileById() {
+  const { userId } = useNumberParams();
+
+  const { data: profileData } = useGetProfileById(userId);
+
+  return <ProfileDetailSection profileData={profileData} />;
+}
