+++ src/app/(main)/profile/[userId]/page.tsx
@@ -1,0 +1,25 @@
+import ProfileByIdDetail from './components/ProfileByIdDetail.client';
+import { Keys, getProfileById } from '@/apis/profile';
+import { HydrationProvider } from '@/components/common/Provider';
+import { Suspense } from 'react';
+
+interface PageProps {
+  params: {
+    userId: string;
+  };
+}
+
+export default function page({ params }: PageProps) {
+  const userId = Number(params.userId);
+
+  return (
+    <Suspense>
+      <HydrationProvider
+        queryFn={() => getProfileById(userId)}
+        queryKey={Keys.getProfileById(userId)}
+      >
+        <ProfileByIdDetail />
+      </HydrationProvider>
+    </Suspense>
+  );
+}
