+++ src/app/(main)/profile/components/ProfileDetail.client.tsx
@@ -1,0 +1,9 @@
+'use client';
+import ProfileDetailSection from './ProfileDetailSection';
+import { useGetProfile } from '@/apis/profile';
+
+export default function ProfileDetail() {
+  const { data: profileData } = useGetProfile();
+
+  return <ProfileDetailSection profileData={profileData} />;
+}
